+++ miden-crypto/src/merkle/smt/large_forest/backend/persistent/property_tests.rs
@@ -54,6 +54,7 @@ proptest! {
         }
     }
 
+
     #[test]
     fn get_correct(
         lineage in arbitrary_lineage(),